Arreglando todos los warnings/errors de Controller/Course.php referentes a psalm

+++ frontend/server/src/Controllers/Problemset.php
@@ -31,7 +31,7 @@ class Problemset extends \OmegaUp\Controllers\Controller {
         string $commit,
         string $currentVersion,
         \OmegaUp\DAO\VO\Identities $identity,
-        int $points,
+        float $points,
         int $order_in_contest = 1,
         bool $validateVisibility = true
     ) {
